Initial commit. Add policy documents.
These documents are intended to allow us to cite the reasons why we
implemented certain business rules by linking them to individual parts
of the original policy.

Markdown wasn't really suitable because of the complicated numbering
systems used in the source documents so I had to use AsciiDoc[0]. It's
important that the numbering sequences are consistent between the plain
text and PDF versions.

AsciiDoc is supported by GitHub so it should be a drop-in replacement
for Markdown.

[0] https://asciidoc.org/