Customise the phase banner and header per env
Often when develping an app it's difficult to tell which environment
you're in. It's more of a problem when they look identical.

This change adjusts two things depending on the environemnt:

* the phase banner
* the bottom border of the header

Both environment-specific changes are handled by the new
EnvironmentHelper. When running in production mode the phase banner will
remain blue, but the content and colours are easily customisable in
the helper.