Add induction programme and number of terms fields
The induction record will hold the ECT's induction programme
choice. For now I've set these to the legacy 3: fip, cip and diy.

There's an ongoing discussion in #214 on how we should store these in
the longer term, but for now we'll go with what we know.

The number of terms is a value recorded when a induction programme is
_closed_ (a finish date is set). It contains a integer number detailing
how many terms the inductee has served. This isn't always closely
related to the dates because people work on different schedules and
there are various other factors that have an influence on how long
induction takes.