Interval test refactoring
The previous commit which added envelopment tests broke loads of stuff
because the seeds weren't (aren't) set up to handle it.

This was used as an opportunity to rework the way the tests work, mainly
that there is now a collection of overlap scenarios in
PeriodHelpers::PeriodExamples that can be reused across the period
models.

The reason we can't really get away with a shared context for this is
that there is different setup required and slightly different rules for
each model.

Additionally some of the simple scopes have been switched to SQL which
is more than adequate and requires no objects to be created.