Add NI number querying to the TRS API client
When adding an ECT we intend to ask first for TRN and DoB. Usually this
will yield a result, but when it doesn't we want to give school admins
another attempt with a different attribute they probably have access to.

The follow up question will try to identify the teacher based on TRN and
national insurance number.

Refs #512, /DFE-Digital/teaching-record-system/pull/1598