Directly link all induction periods to teachers
Originally we made induction periods dependant on ECT at school periods
because before the AB portal work came along schools were intended to
report the ECT's AB.

However, as the AB portal work jumped the queue we had to add a second
relationship that directly links induction periods with teachers,
meaning there were two routes between AB and teacher.

This small change made lots of the code we added more complicated and
simple things, like validating that a teacher's induction periods don't
overlap, arduous.

So, this change simplifies it by removing the link via ECT at school
period. It means later we might need to add more logic elsewhere, like
when a teacher leaves a school but has an ongoing induction period, we
need to deal with it appropriately.

But, for now, this greatly simplifies things.