Fix pass and fail recording for appropriate bodies
This was broken by the change in #790 that limits visible ECTs to those
who are owned by the AB, and both passing and failing ECTs means they're
no longer owned by the 'Success' page, so it breaks.

The tests didn't pick that up because they were mocking the actual
closure of the induction period, so that's been removed and the actual
final induction period is checked.