Make sign_in_with_xxx methods return sessions user
Before this change the methods returned nil which made it difficult to
ensure events were attributed to the right author. We won't have the
actual user session object but the attributes will be the same, so we
can compare them:

  expect(AppropriateBodies::ClaimAnECT::RegisterECT).to have_received(:new).with(
    appropriate_body:,
    pending_induction_submission:,
    author: instance_of(Sessions::Users::AppropriateBodyUser)
  ) do |arguments|
    arguments.fetch(:author).tap do |author|
      expect(author.name).to eql(user.name)
      expect(author.email).to eql(user.email)
    end
  end

I resisted implementing a custom `def  ==(other)` because that might
have unintended consequences elsewhere, perhaps worth exploring.