Record events for ECT registration by an AB
The register ECT class now accepts an author: argument. The author must
be a Sessions::User or the operation will fail with NotASessionsUser.

The two events that might occur here are:

* appropriate_body_claims_teacher
* teacher_name_updated_by_trs

They are in this PR to show the intent of how we might record events and
give an opportunity for feedback, but I think this approach is sound and
I prefer being explicit over trying to hook things in automatically or
via callbacks.

The Events::Record is currently generic and could be used for all
events, or we could create small specialised classes for different
types. This might make more sense in the long run as the event type text
will be in its own class so (hopefully) easy to find/change.