Stop automatically updating Playwright in bin/setup
This causes unexpected changes to creep into commits. Instead we run a
check before running the test suite that fails when there's a major
version mismatch between Playwright JavaScript and Ruby versions.

This should only fail in Dependabot pull requests that bump either the
NPM package or Gem. We'll have to update the corresponding version
manually within the Dependabot PR.