Early exit when claimed on check instead of find
Now we want to allow any AB to see ECTs without claiming them we want to
nudge the 'already claimed by another AB' early exit back a
step. Instead of it happening when finding the ECT it now happens if the
AB tries to begin a claim process on them.

We'll hide the claim button too so this is more of a safeguard (i.e., if
two ABs are trying to claim the same ECT simultaneously)