Move the claim ECT button to the top of check page
Previously we had the confirm details and continue buttons at the bottom
of the page but now we're changing the journey from just claim to find
and claim, we want to move it up. This is because the 'check' page
contains far more data and the button would be pushed way below the
fold.

There are no form elements on this page any more so the error summary
has been removed too.

Fixes #280, #979