Refactor and simplify the primary nav
We don't (yet) need to know anything about the user in order to render
the nav. We're already guarding against the wrong kind of user accessing
the page at the controller level so it doesn't make sense to do it here
too.

Now we use the path prefix to determine the 'area' and then use that
area value to ensure we're:

* setting the right serivce_url
* adding the right nav items

When no area is matched, i.e., when the user's on the landing page,
they'll be assigned :none and nothing is rendered.

The current page behaviour is already implemented in the service nav
component[0] so we don't need to pass current (or active) in, plus we
can pass in the whole set of nav items as an argument so don't need to
loop.

[0] https://govuk-components.netlify.app/components/service-navigation/#input-slim-with-navigation-items