Add builder class for pending induction submissions
When we're closing an induction period we populate a
PendingInductionSubmission record and use it to close the
InductionPeriod. We do this because soon we'll add bulk editing (via CSV
uploads/API) and this allows us to do everything in a consistent manner.

We want to ensure the finished_on date is later than the started_on
date, so this change adds a PendingInductionSubmissions::Build class
that lets use use the InductionPeriod (which will soon be closed) to
initalize the PendingInductionSubmissions, so we can compare the two
dates.