Fix bug where induction period not saved
This happens becuase we were checking for the truthiness of each success
step but calling `.create` on a invalid model returns the unsaved record
which is truthy.

Now we call `#persisted?` at the end of the method and only call
record_induction_period_event if the induction period was saved.