Bump finished_on by one day if matches started_on
I don't think we should allow any induction periods that start and
finish on the same day.

If they were added in error they should be deleted instead.

To support this and prevent us from writing invalid records to the
database during the import, bump the finished_on by a day if it matches
the started_on