Remove the unique index on la code and establishment no
We do need these to be unique when it comes to bulk uploading (if we
continue to do it the same way as it used to work in the AB portal) but
the underlying DQT data actually has a handful clashes.

For now, allow them in and we'll need to fix them later - if we do
decide on using this combo as unique identifiers.