Add class for searching admin emails case insensitively
We are not using Arel's .matches here because `ilike` adds some danger
in that people could add wildcards to searches.

Instead convert both the email address column and search term to lower
case. The users table is not likely to be around too much longer and
will never be big, so this approach is fine.

This fixes a bug where people need to match the email exactly to be able
to log in, and they won't have set it up themselves so don't know for
certain what they need to do.