Simplify Teachers::Search
This change reworks the teachers search class and makes it more
generic, and removes its need to have knowledge about where it's being
called for.

It also makes the critiera combinable, so we'll be able to support
searching within an AB's ECTs by providing both a search term and an
appropriate body.

It also will allow us to search by multiple appropriate bodies if
needed.