Make statuses depeendant on induction period presence
We ran into a bug in staging where teachers with no induction periods
are marked as 'Induction paused' because they have no ongoing induction
periods either.

This shouldn't happen in production yet but we can still fix the bug by
only overriding the TRS induction status when we have both a teacher
record and it has induction periods.