Fix the induction event happened_at values
I implemented this wrong first time around because I was thinking of the
event from the point of view of when it was recorded, not from the
actual final timeline.

On the timeline we want the date the person started or finished an
induction period, not the date the AB informed us.

We'll need to fix some records in prod but that is thankfully
straightfoward.