Disable MIGRATION settings in the .env
It's likely a new dev who sets up the app won't want to dive straight
into migration stuff, so disable this to start with.

This also has the benefit of allowing the tests/app to run right after
copying the .env.example to .env and the lack of a database URL won't
cause problems.