Fix `db:prepare` rake task when ran without database
When setting up the app from scratch we run `bin/setup`, which calls
`db:preapre`. If a database has not been created yet, it will attempt to run
the `setup_search_configuration` enhancement (which is invoked before
`db:prepare`) and it will fail as it modifies the database.

The `db:prepare` command will succeed if `db:create` is ran first. We have a
bit of a chicken/egg scenario here and its complicated by the fact `db:prepare`
does not invoke `db:create` enhancements (so we can't just enhance `db:create`
with `db:setup_search_configuration` in a block to ensure it runs after).

I can't see a nice way around this; the best I could come up with is to modify
the `setup_search_configuration` rake task to create the database if it doesn't
yet exist. We can also enhance `db:create` with a block to setup the search
config after the database is created.

This means we can run `db:preapre` when the database both exists and doesn't
yet exist and `db:create` independently. We should be fine to drop
`db:schema:load` as it will be covered by one of the other two scenarios.