Fix rake tasks failing in review app
For some reason the review apps have started failing since changing the rake
task search config setup ordering. We're not entirely sure what the issue is,
but it seems to be enough to have the safety net of creating the DB if it
doesn't exist in the search config rake task.

Revert the enhancements to the ones that worked. Retain the DB creation if not
exist.