Add scope for ordering by refresh date nulls first
PostgreSQL's order clause defaults to 'NULLS LAST' when sorting
ascending ('ASC').

We want to find rows in ascending order so the ones refreshed longest
ago are first - but we also want rows that have never been refreshed
(are 'NULL') to be first.

To do that we need to specify the where nulls go as part of the
query. It's not supported in Rails by default so we can use Arel.

https://www.postgresql.org/docs/current/queries-order.html