Add ERD PNG diagram to README
This gem uses rails-erd (https://voormedia.github.io/rails-erd/) to generate
an "entity relationship diagram", also known as a ERD diagram.

Having it in the README is useful for new starters on the project. The
diagram can be manually regenerated, but by default is regenerated
whenever you run `rails db:migrate`. This hopefully will keep the
diagram and the database in sync.