Add script that deletes 1 day periods
We had a problem with imported data where some records were adjusted so they
had a length of 1 day instead of 0 (no longer allowed). That adjustment meant
there's a chance they overlap other periods.

We want to delete those periods that:
  - start on the same day as another period
  - are 1 day long

The included query:
  - finds all teachers that have 2 induction periods starting
    on the same day
  - finds induction periods belonging to those
    teachers
  - limits those induction periods to ones that are 1 day long

Rather than execute the deletion SQL directly we'll just hardcode the
IDs of the 100 offending records.