Fix flaky spec by setting `number_of_terms` to 1 in factory
The factory sets a `random number_of_terms` between 1 and 6, which sometimes
matches the expected value of 4 in the test. This caused the `.to change` matcher
to fail since the value didn't actually change.

There is no value in setting this number dynamically, we are setting it to 1 as that is the most common `number_of_terms`