Add `training_period` references to `LeadProviderActivePeriod` and `SchoolPartnership`
As per the UML diagram, we want `TrainingPeriod` to have a zero/one
relationship to `SchoolPartnership` and `LeadProviderActivePeriod` (originally
`ActiveLeadProvider` in the UML). Both `LeadProviderActivePeriod` and
`SchoolPartnership` can have many training periods.

I've opted for the `TrainingPeriod` -> `LeadProviderActivePeriod` relationship
being an `expression_of_interest` and the `TrainingPeriod` ->
`SchoolPartnership` relationship as a `confirmed_school_partnership`.

There can only be one or the other, not both (and there must be at least one).