Ensure `SchoolPartnership` school is consistent with school periods
We link a `TrainingPeriod` to a school already via one of the school periods.
When connecting a `TrainingPeriod` to a `SchoolPartnership` we need to ensure
the `school` of the partnership is consistent with the school periods.

We need to duplicate this relationship with a `School` due to the timings of
how partnerships and training periods can be created (either one could come
first and needs a reference to the school, and they can both exist
independently).