Add event recording for mentorships
When mentorships are created we create two events, one for the mentor
and one for the mentee. This will make the 'ownership' of the event more
straightforward too.

For example, if the ECT is at School A and the mentor at School B, and
we look at the events for School A's POV, we'll see one of ECTs began
being mentored while from School B's POV we'll see one of their mentors
began mentoring.