Add mark_teacher_as_deactivated! to Teachers::Manage
This method allows us to set a teacher as having been deactivated in
TRS. This means we need to take further action at our end to reconcile
the differences.

Usually it's a result of a teacher having more than one
record (and TRN) so the records are merged, and this is the one that's
left over.

We won't know what happened on the TRS side yet and it's not been
surfaced in the API, so just marking for now.