Use JSON args format
Using the script style flags warnings from Docker lint because the
argument format means OS singnals can't be detected.

> When you use shell form, the executable runs as a child process to a
> shell, which doesn't pass signals. This means that the program running
> in the container can't detect OS signals like SIGTERM and SIGKILL and
> respond to them correctly.

https://docs.docker.com/reference/build-checks/json-args-recommended/

This means we're hardcoding the port, but it's only in two places a
couple of lines apart so it should be obvious. I can't make it use
${PORT} because it's the shell (not Docker) that expands it and the
busybox shell just isn't picking it up regardless of how I try to pass
the value in (tried echoing, passing off to a separate script)