Organise seeds by individual files
The seeds file is getting very long and increasingly difficult to change.

We want to look at splitting the seeds out by type so its easier to index into
and add to them.

We probably want to do another pass at this as there's a lot of
inter-dependencies between the seeds just now as well.