Move TRS::FakeAPIClient to lib
When it's required like this in the middle of a class it's both ugly and
it seems to not reload properly after the app has crashed.

Now it's going to be moved in other envs it makes sense to move it to
lib. We can add some tests there too.