Fix `unpermitted_params` warning from pagination concern
In the pagination concern we use `params.permit` to retrieve the `page` and
`per_page`, however these parameters are often in addition to other parameters
(such as filters) and result in a warning in the Rails logs.

Use `params.dig` instead of `params.permit` to avoid warnings when paginating
and filtering at the same time.