Disable reading from TRS in sandbox environment
We do not want to allow syncs/data pulls from TRS in the sandbox environment to
overwrite data in the database (as this may end up confusing lead providers).

The main source of this is in the recurring jobs which are only enabled in
production, however there are two jobs `FailECTInductionJob` and
`PassECTInductionJob` which both write data to TRS and then read it back via
the `Teachers::RefreshTRSAttributes` service.

Introduce a `trs_write_only` config variable which is `false` by default and
`true` in sandbox to prevent the `Teachers::RefreshTRSAttributes` service from
making any changes.