Add seed data services for sandbox
We want to be able to generate seed data in sandbox.

The seed data initially will be registration periods, lead providers and
statements. The statements data should be close to what we have in production.

We want the data to remain consistent if we run the seeds multiple times (we
should end up with the same registration periods, lead providers and
statements).

Add sandbox seed data services. A `Base` service is used to centralise logging
and environment checks.

Add a rake task for convenience; `bundle exec rake sandbox_seed_data:generate`.