Fix flakey statements sandbox seed spec
When generating statements for sandbox seed data we only get a `payable`
statement if the random payment date includes the current date.

This has a chance of happening for a number of statements in the seeds, but
isn't guaranteed to and so it causes the test to ensure we generate statements
with all states to be flakey.

Instead, we adjust the window for statements with dates near the current date
to ensure they always end up as `payable`.