Fix bug in page_data where arguments are ignored
There was a bug in #page_data where if header was false it would return
early, before the caption, backlinks, breadcrumbs etc had been dealt
with.

Originally the method was much simpler but extra functionality has been
added and now the return doesn't make sense - sometimes we might want a
page with a manually-set header to have a back link for example.

Now, the default value for header is :use_title instead of nil. This
means we don't need to worry about differentiating between nil and false
and can be more explicit.

I've also rewritten the tests to cover all of the functionality properly
and no longer return a hash of values (which originally were there to
make testing easier) in favour of having the tests pull data out of
content_for to ensure it's set correctly.