Make the email page more accessible
On the ECT registration flow the label wasn't really describing the
question, it was more of a hint.

I've made the page heading the label and set the old label text as the
hint.

With this being fixed we can also change the features to use the label
text when selecting the input.