Add `ParityCheck::TokenProvider`
When running the parity check we need to be able to grab API tokens for each
lead provider in order to make valid requests.

We also need to generate tokens for the lead providers after refreshing the
mgiration database (so that we have shared tokens in migration between ECF and
RECT and can therefore make the same requests to both services).