Add `token` method to `TokenProvider`
We need to be able to retrieve a token for a given lead provider as part of the
parity check in order to make valid API calls.

We could have encapsulated this in the `TokenManager` service, but as we will
want to cleanly remove the parity check related code after we are finished
migrating across it makes sense to house it in the
`ParityCheck::TokenProvider`.