Add `ParityCheck::Client` and associated models/services
We want to abstract the complexities involved with making parity check requests
into a `Client`. The client will be responsible for taking an endpoint
definition, constructing valid requests, making the requests and returning the
results along with request duration information.

Add `Endpoint` model to describe a callable endpoint.

Add `RequestBuilder` service to take an `Endpoint` and construct the different
aspects of a request (path, headers, etc).

Add `Client` to perform the requests with `HTTParty`. To begin with we will
only support basic GET endpoints.