Add expression of interest relationship
The EOI represents the preference shown by a school to work with a lead
provider. It's explicitly named 'expression_of_interest' in both the
database and the models to avoid confusion between it and the **actual**
lead provider that will be linked to a traiing period (via the school
partnership).