Make `Endpoint` an `ActiveRecord` model
Initially I thought `Endpoint` could just be defined and loaded from a YAML
file, but in reality this creates more work as the job will accept endpoints
and then we will need to re-look them up from the method/path when we come to
process each request.

Instead, its going to make more sense to have the `Endpoint` as an
`ActiveRecord` model that is associated with each `Request` and seeded into the
database.