Validate presence of either EOI or school partnership
A training period can exist in one of two 'states', an expression of
interest and a confirmed training period.

Therefore, it needs to at least be related to either the active lead
provider or school partnership. For records that begin as an EOI and
progress to a confirmed training period we'll hold both (but the school
partnership takes precedence)