Support POST/PUT and query parameters
We want to update the parity check to support POST and PUT endpoints, as well
as endpoints that contain query parameters.

Update `RequestBuilder` to to extract query parameters and the request body
from the request options.

Pass the body and query parameters through to `HTTParty` in the parity check
`Client`.

Contains an example request body for now for the tests; we can change this when
we add the first set of POST/PUT endpoints.