Switch `ParityCheck::Client` to use Faraday
We are switching the parity check to use Faraday to be consistent with
other areas of the app.

It makes sense to take advantage of the `in_parallel` support Faraday
offers to make the parity check requests in parallel (to both services at the
same time).  This should make the overall parity check run quicker.

Extracts the response benchmarking to a Faraday middleware to make the
client cleaner.