Add state machines to `ParityCheck::Run` and `ParityCheck::Request`
Add state machines to `Run` and `Request` in an effort to clean up how they
progress through the parity check pipeline.

A `Run` can be `pending`, `in_progress` or `completed` (note there is no
`queued` state as we don't want a long running task sitting in a job.

A `Request` can be `pending`, `queued`, `in_progress` or `completed`. We have a
`queued` state here to describe the state e request is in after its been queued
but not yet picked up by a worker.

Fleshes out the factories to include states and also some convenience traits
for request methods.

Add `mode` attribute to `Run` so that we can determine if it should be
concurrent or sequential.