Update `RequestHandler` to call `RequestDispatcher`
When a request has been handled, it may be the last request in the batch. If it
is we need to kick off the next set of requests.

Call the `RequestDispatcher` when a request finishes. The dispatcher is
thread-safe and if other requests are still on-going it won't do anything.

Update the request state during handling.