Support multiple endpoints with the same method/path
At the moment if we define two endpoints with the same path under the same
request method in the YAML file we only get one after parsing the file due to
it being structured in a hash.

To avoid this, we can change the endpoints to be array items; this will allow
duplicates. The use case for this is when testing, for example, an index action
once with pagination and a separate time with filters.