Add `group_name` and `description` to `Endpoint`
When we come to render an endpoint in a view we want to be able to describe the
endpoint and group them.

Add `description` which will be used for the checkbox labels.

Add `group_name` which will allow us to group endpoints sensibly in the UI.

Add validation to ensure endpoint paths do not contain query parameters (these
should be specified under the `options[:query]` in the YAML file.