Add admin page to start a parity check
Add the initial page to start a parity check; this is only accessible to
administrators and only if the environment supports parity checks.

The `Runner` has been updated so we can use it as a form object.

If no endpoints are selected the user will see an error message.

On submitting a run we show a flash notice to inform the user it has been
started.

Endpoints are grouped to make selection easier.