Add progress methods to `ParityCheck::Run`
We want to be able to determine the progress of a parity check run and an
estimate of when it will complete so that we can surface these details in the
UI.

Calculate `progress` based on the number of completed requests out of the total
requests (returning a whole number percentage).

Calculate the `estimated_completion_at` by determining the total runtime for
the current progress and inferring a completion time based on the amount of
progress remaining.