Add in-progress and pending parity checks to start page
We want to surface the in-progress and pending parity checks on the start page,
so that the user can clearly see the progress of in-flight checks.

Display the in-progress parity check, including a progress bar and percentage
complete.

Display a list of pending parity checks and an indication of when the next
pending check will be picked up.