Replace output_fee boolean with fee_type
This change is an attempt to make statement fee types a bit clearer. In
ECF1 the logic was that if `output_fee == true` the statement was an
output fee, and if `output_fee == false` it's a service fee.

It's not intuitive or searchable, and given the schema it'd be
impossible for someone to work out what a service fee is.

This change replaces the output_fee column with fee_type, which is an
enumerated type where the options are `service` or `output`.