Add completed parity checks to seed data
This makes developing the parity check locally easier if we can have completed
parity checks planted.

I've not planted any pending/in-progress runs as the would just be instantly
picked up by the job workers and become completed when the app boots anyway -
we can easily kick these off manually.

Only plant on environments the parity check is enabeld on.