Remove the 'not null' from trs first and last names
TRS doesn't enforce the presence of the teacher's first and last names
but we do. This means we have sync errors when refreshing teacher
details.

This change removes the 'not null' constraint from TRS first and last
name. It needs to drop/recreate the search column because PostgreSQL
won't allow column changes when they're depended on by other fields.

It also removes the 80 character limit. We don't enter this data
ourselves and are reliant on TRS, so just accept what we're given.