Add convenience methods/scopes to parity check models
Add methods so that we can pull out key metrics at both the `Run`, `Request`
and `Response` level (namely `match_rate` and `rect_performance_gain_ratio`).
Note that these methods may appear to return differing results at each level
due to the way we are averaging the averages.

Add `has_many, through:` for easily grabbing endpoints and lead providers for a
run; we distinct and sort the result by default for ordering in the UI.