Add page to view a completed parity check
Add page to view a completed parity check. We can click through from the list
of completed parity checks to get to the page.

There is a high-level description of the completed run and then a breakdown of
each requests by lead provider.

Rename the parity check feature specs for clarity.