Ensure domain-model.png is kept up to date
We often find that the `domain-model.png` file is out of sync with the actual
domain model according to the latest migrations.

I think this can be due to either forgetting to commit the updated diagram or
if graphviz isn't installed locally it won't regenerate the diagram.

Add a check in CI to ensure the diagram committed is consistent with the
migrations.