Show a parity check response in detail
We want to be able to click into a response that is different in order to see
a diff of the response bodies and identify what's changed.

Add routes/actions for showing a parity check response.

Add `diffy` to generate diffs of the response bodies.

Add more realistic JSON bodies to the seed data.

Add convenience methods to models and view helper methods.