Add School::LatestRegistrationChoices service
This service is intended to allow us to populate a screen on the ECT
registration journey where we show the school's

* most recently chosen appropriate body
* most recently chosen lead provider and delivery partner

The appropriate body choice is easier, currently all ABs in the service
are considered 'active'.

Selecting the lead provider and delivery partner is more complex because
lead providers may or may not be active in the current registration
period, and they may or may not be paired with the same delivery partner
as they were last time.

So, to pick them we need to:

* ensure the lead provider is active in the registration period
* order properly, so when the lead provider is active but paired with
  more than one delivery partner, we want to display the oldest one as
  we deemed it's most likely their _default_ choice)