Add `ParityCheck::Filter::ResponseBody` model
This model will be used in the form to filter a response body diff by the
available JSON keys.

It will accept a `response` and `selected_key_paths` submitted from the form
(via checkboxes).

It exposes a `filterable_key_hash` which will be used to generate the
checkboxes in a tree structure.

It accepts `selected_key_paths` in the format that they will natively be
submitted by the browser; for an array set as an `input` value this is a
space-separated list, for example the key path `[:key, :nested]` it would
submit as `key nested`.

The `filtered_response` method applies the selected filter to the response
bodies (excluding fields that are not selected). By default all fields will be
treated as selected (if `selected_key_paths` is `nil`).