Add helper to render checkboxes for a `filterable_key_hash`
This helper will accept a `filterable_key_hash` from a `ResponseBody` filter
and output it as a nested HTML list of checkboxes.

Note the structure of the list nesting is slightly different to what you might
expected; the nested `ul` is contained with the parent `li`, so :

```
<ul>
  <li>Parent
    <ul>
      <li>Child</li>
    </ul>
  </li>
<ul>
```

Instead of:

```
<ul>
  <li>Parent</li>
  <li>
    <ul>
      <li>Child</li>
    </ul>
  </li>
</ul>
```

It is like this to make the subsequent styling more straight forward (get show
connections between the keys).