Add Stimulus controller for response body filtering
When filtering the response body the user will be interacting with checkboxes
in a tree structure matching the JSON keys.

If they check a deeply nested checkbox, we want to also check all checkboxes
that preceed it in the tree.

If they uncheck a top-level checkbox we want to uncheck all its direct
children.

Add `ResponseBodyFilterController` and register it with Stimulus.

Import controllers into `application.js`.